fixed intinite recursion when setting Player.tournament
